Move `CurtzStoreManager`, `CurtzTokenService` to production
