+++ Tests/Curtz/Token/CurtzTokenStoreUnitTests.swift
@@ -8,97 +8,9 @@
 import XCTest
 import Curtz
 
-/*
- 1. Store access_token and refresh_token
- 2. Delete access_token and refresh_token
- 3. Retrieve access_token and refresh_token
- */
-
 
 protocol StoreQueryable {
     var query: [String: AnyObject] { get }
-}
-
-
-
-
-final class CurtzStoreManager: StoreManager {
-   
-  
-    private let store: Store
-    
-    init(with store: Store){
-        self.store = store
-    }
-    
-    func save(_ val: String, forKey key: String, completion: @escaping (SaveResult) -> Void) {
-        store.add(val, key: key) {[weak self] result  in
-            guard self != nil else { return }
-            switch result  {
-            case let .failure(error):
-                switch error {
-                case .failedToSave:
-                    completion(.failure(.failedToSave))
-                default:
-                    completion(.failure(.general(error)))
-                }
-            default:
-                completion(.success(()))
-            }
-        }
-    }
-    
-    func retrieveValue(forKey key: String, completion: @escaping(RetrieveResult) -> Void){
-        store.search(forKey: key) {[weak self] result in
-            guard self != nil else { return }
-            switch result {
-            case let .failure(error):
-                switch error {
-                case .notFound:
-                    completion(.failure(.notFound))
-                default:
-                    completion(.failure(.general(error)))
-                }
-            case let .success(retrievedValue):
-                completion(.success(retrievedValue))
-            }
-        }
-    }
-    
-    func update(_ val: String, forKey key: String, completion: @escaping(UpdateResult) -> Void) {
-        store.update(val, forKey: key) {[weak self] result in
-            guard self != nil else { return }
-            switch result {
-            case let .failure(error):
-                switch error {
-                case .failedToUpdate:
-                    completion(.failure(.failedToUpdate))
-                default:
-                    completion(.failure(.general(error)))
-                }
-            case .success:
-                completion(.success(()))
-            }
-        }
-    }
-    
-    
-    func removeValue(forKey key: String, completion: @escaping(DeleteResult) -> Void){
-        store.deleteValue(key: key) {[weak self] result in
-            guard self != nil else { return }
-            switch result {
-            case let .failure(error):
-                switch error {
-                case .failedToDelete:
-                    completion(.failure(.failedToRemove))
-                default:
-                    completion(.failure(.general(error)))
-                }
-            case .success:
-                completion(.success(()))
-            }
-        }
-    }
 }
 
 final class CurtzStoreManagerUnitTests: XCTestCase {
